feat: 学生列表分页查询

+++ pojo/src/main/java/com/wangwei/entity/ClassTeacherRelation.java
@@ -9,7 +9,7 @@ import java.io.Serializable;
 @Builder
 public class ClassTeacherRelation implements Serializable {
     // 班级ID
-    private Integer classId;
+    private Long classId;
     // 教师ID
-    private Integer teacherId;
+    private Long teacherId;
 }
